Move settings screens to new auth tab, replacing profile tab; create empty settings tab
