+++ frontend/navigation/SettingsContainer.js
@@ -1,8 +1,6 @@
 import * as React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { NavigationContainer } from '@react-navigation/native';
-import LoginScreen from '../screens/LoginScreen';
-import SignupScreen from '../screens/SignupScreen';
 import SettingsScreen from '../screens/SettingsScreen';
 
 
@@ -16,8 +14,6 @@ function SettingsContainer() {
         screenOptions={{headerShown: true}}
       >
         <Stack.Screen name="Settings" component={SettingsScreen} />
-        <Stack.Screen name="Login" component={LoginScreen} options={{ title: 'Login' }} />
-        <Stack.Screen name="Signup" component={SignupScreen} options={{ title: 'Signup' }} />
       </Stack.Navigator>
     </NavigationContainer>
   );
